Move Zendesk adapter code to pkg/adapter and cmd/
