+++ cmd/zendesksource-adapter/kodata/HEAD
@@ -1,0 +1,1 @@
+../../../.git/HEAD
