+++ config/500-controller.yaml
@@ -57,6 +57,8 @@ spec:
         # Source adapters
         - name: SLACKSOURCEADAPTER_IMAGE
           value: ko://github.com/triggermesh/knative-sources/cmd/slacksource-adapter
+        - name: ZENDESKSOURCEADAPTER_IMAGE
+          value: ko://github.com/triggermesh/knative-sources/cmd/zendesksource-adapter
 
         securityContext:
           allowPrivilegeEscalation: false
