+++ pkg/adapter/zendesksource/adapter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package adapter
+package zendesksource
 
 import (
 	"context"
@@ -33,8 +33,8 @@ type zendeskAdapter struct {
 	logger  *zap.SugaredLogger
 }
 
-// New adapter implementation
-func New(ctx context.Context, aEnv adapter.EnvConfigAccessor, ceClient cloudevents.Client) adapter.Adapter {
+// NewAdapter adapter implementation
+func NewAdapter(ctx context.Context, aEnv adapter.EnvConfigAccessor, ceClient cloudevents.Client) adapter.Adapter {
 	env := aEnv.(*envAccessor)
 	logger := logging.FromContext(ctx)
 	eventsource := env.Subdomain + "." + env.Name
