+++ pkg/adapter/zendesksource/env.go
@@ -1,9 +1,12 @@
 /*
 Copyright (c) 2020 TriggerMesh Inc.
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-   http://www.apache.org/licenses/LICENSE-2.0
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -11,14 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package adapter
+package zendesksource
 
 import (
 	"knative.dev/eventing/pkg/adapter/v2"
 )
 
-// EnvAccesor for configuration parameters
-func EnvAccesor() adapter.EnvConfigAccessor {
+// EnvAccessor for configuration parameters
+func EnvAccessor() adapter.EnvConfigAccessor {
 	return &envAccessor{}
 }
