+++ zendesk/cloudbuild.yaml
@@ -2,17 +2,6 @@ steps:
 
 - name: 'gcr.io/kaniko-project/executor:latest'
   args:
-  - --dockerfile=${_KSOURCE}/cmd/adapter/Dockerfile
-  - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-adapter:${COMMIT_SHA}
-  - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-adapter:${_KANIKO_IMAGE_TAG}
-  - --cache-repo=gcr.io/$PROJECT_ID/${_KSOURCE}-source-adapter/cache
-  - --cache=${_KANIKO_USE_BUILD_CACHE}
-  - --no-push=${_KANIKO_NO_PUSH}
-  - ${_KANIKO_EXTRA_ARGS}
-  waitFor: ['-']
-
-- name: 'gcr.io/kaniko-project/executor:latest'
-  args:
   - --dockerfile=${_KSOURCE}/cmd/controller/Dockerfile
   - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-controller:${COMMIT_SHA}
   - --destination=gcr.io/$PROJECT_ID/${_KSOURCE}-source-controller:${_KANIKO_IMAGE_TAG}
